gui/video: Use asyncio to display gui
Using the async functionality of `server/camera`, the GUI can also wait
for the frame to be delivered, before drawing. During which, the
`Camera` class can receive more images.

The images received must be processed as `dearpygui` sends them to the
GPU before displaying. So, they must first be converted from BGR to RGB
and then to a 1D array. Then to 32 bit floats.

Added a new async method: `update()` which updates the raw texture on call
`start_cmd()` cannot be called more than once.